+++ addOns/automation/CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
+
+
+## [0.34.0] - 2023-11-16
 ### Added
 - Show column control in the Automation tab to allow to show/hide columns and auto resize them (`Pack All Columns`).
 
@@ -320,6 +323,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - First version.
 
+[0.34.0]: https://github.com/zaproxy/zap-extensions/releases/automation-v0.34.0
 [0.33.0]: https://github.com/zaproxy/zap-extensions/releases/automation-v0.33.0
 [0.32.0]: https://github.com/zaproxy/zap-extensions/releases/automation-v0.32.0
 [0.31.0]: https://github.com/zaproxy/zap-extensions/releases/automation-v0.31.0