+++ addOns/commonlib/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Add utilities for time-based checks, migrated from Active scanner rules add-on.
+
 ### Changed
 - Added solution to 'Fingerprinting' (Issue 8056).